+++ flowman-core/src/main/scala/com/dimajix/flowman/jdbc/OracleDialect.scala
@@ -1,5 +1,5 @@
 /*
- * Copyright 2022 Kaya Kupferschmidt
+ * Copyright 2022-2023 Kaya Kupferschmidt
  *
  * Licensed under the Apache License, Version 2.0 (the "License");
  * you may not use this file except in compliance with the License.
@@ -258,12 +258,12 @@ class OracleCommands(dialect: BaseDialect) extends BaseCommands(dialect) {
         super.addColumn(statement, table, columnName, dataType, isNullable, charset, collation, comment)
         comment.foreach { c =>
             val sql = dialect.statement.updateColumnComment(table, columnName, dataType, isNullable, charset, collation, comment)
-            statement.executeUpdate(sql)
+            JdbcUtils.executeUpdate(statement, sql)
         }
     }
 
     override def updateColumnComment(statement:Statement, table: TableIdentifier, columnName:String, dataType: String, isNullable: Boolean, charset:Option[String]=None, collation:Option[String]=None, comment:Option[String]=None) : Unit = {
         val sql = dialect.statement.updateColumnComment(table, columnName, dataType, isNullable, charset, collation, comment)
-        statement.executeUpdate(sql)
+        JdbcUtils.executeUpdate(statement, sql)
     }
 }