+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * github-314: Move avro related functionality into separate plugin
 * github-307: Describe Flowmans security policy in SECURITY.md
 * github-315: Create build profile for CDP 7.1 with Spark 3.3
+* github-317: Perform retry on failing JDBC commands
 
 
 # Version 0.30.0 - 2023-01-03